requirements.txt
python-dotenv
flask
flask-cors
google-genai
markitdown[all]
groq
faster-whisper
watchdog
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
chromadb
sentence-transformers
torch
tree-sitter
tree_sitter_language_pack